requirements.txt
numpy
torch
scikit-learn
matplotlib
pyyaml
norse
